# services/tracking-gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API for the WR any% automatic route tracking application.\nWebSocket tracker: ws://host/v1/ws/tracker — see docs/contracts/websocket-tracker-protocol.md",
        "title": "WR any% API",
        "contact": {},
        "version": "0.1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/v1/routes": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "routes"
                ],
                "summary": "List routes for the authenticated user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by device UUID",
                        "name": "device_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Max results per page (default 50, max 200)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Pagination cursor from previous response",
                        "name": "cursor",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.RouteListEnv"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/http.ApiError"
                        }
                    }
                }
            }
        },
        "/v1/routes/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "routes"
                ],
                "summary": "Get a single route by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Route UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.RouteEnv"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/http.ApiError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.ApiError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Removes the route and its trip-match links. The underlying trips are preserved.",
                "tags": [
                    "routes"
                ],
                "summary": "Delete a route",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Route UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/http.ApiError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.ApiError"
                        }
                    }
                }
            }
        },
        "/v1/routes/{id}/points": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "routes"
                ],
                "summary": "Get template polyline points of a route",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Route UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.RoutePointsEnv"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/http.ApiError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.ApiError"
                        }
                    }
                }
            }
        },
        "/v1/routes/{id}/trips": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "routes"
                ],
                "summary": "List trips attached to a route",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Route UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Max results per page (default 50, max 200)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Pagination cursor from previous response",
                        "name": "cursor",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.RouteTripListEnv"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/http.ApiError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/http.ApiError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.ApiError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "http.ApiError": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string"
                },
                "error": {
                    "type": "string",
                    "example": "unauthorized"
                }
            }
        },
...
        "http.RouteEnv": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/http.RouteItem"
                },
                "error": {
                    "type": "string"
                }
            }
        },
        "http.RouteItem": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "distance_m": {
                    "type": "number"
                },
                "end_lat": {
                    "type": "number"
                },
                "end_lon": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "start_lat": {
                    "type": "number"
                },
                "start_lon": {
                    "type": "number"
                },
                "status": {
                    "type": "string"
                },
                "trips_count": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "http.RouteListEnv": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/http.RouteListResponse"
                },
                "error": {
                    "type": "string"
                }
            }
        },
        "http.RouteListResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/http.RouteItem"
                    }
                },
                "next_cursor": {
                    "type": "string"
                }
            }
        },
        "http.RoutePointItem": {
            "type": "object",
            "properties": {
                "lat": {
                    "type": "number"
                },
                "lon": {
                    "type": "number"
                }
            }
        },
        "http.RoutePointsEnv": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/http.RoutePointItem"
                    }
                },
                "error": {
                    "type": "string"
                }
            }
        },
...
        "http.RouteTripItem": {
            "type": "object",
            "properties": {
                "distance_m": {
                    "type": "number"
                },
                "duration_sec": {
                    "type": "integer"
                },
                "ended_at": {
                    "type": "string"
                },
                "match_score": {
                    "type": "number"
                },
                "matched_at": {
                    "type": "string"
                },
                "started_at": {
                    "type": "string"
                },
                "trip_id": {
                    "type": "string"
                }
            }
        },
        "http.RouteTripListEnv": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/http.RouteTripListResponse"
                },
                "error": {
                    "type": "string"
                }
            }
        },
        "http.RouteTripListResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/http.RouteTripItem"
                    }
                },
                "next_cursor": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer JWT access token. For WebSocket upgrade ?access_token=\u003cjwt\u003e is also accepted.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}